pyproject.toml
[project]
name = "projeto-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.12.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "supabase>=2.22.2",
]
